+++ packages/cdn/react/mui-react.js
@@ -977,15 +977,15 @@ var Button = function (_React$Component) {
 
       if (ev.type === 'touchstart' && ev.touches) clickEv = ev.touches[0];else clickEv = ev;
 
-      // choose diameter
-      var diameter = Math.sqrt(offset.width * offset.width + offset.height * offset.height) * 2;
+      // calculate radius
+      var radius = Math.sqrt(offset.width * offset.width + offset.height * offset.height);
 
       // add ripple to state
       this.setState({
         ripple: {
-          xPos: Math.round(clickEv.pageX - offset.left),
-          yPos: Math.round(clickEv.pageY - offset.top),
-          diameter: diameter
+          top: Math.round(clickEv.pageY - offset.top - radius) + 'px',
+          left: Math.round(clickEv.pageX - offset.left - radius) + 'px',
+          diameter: radius * 2 + 'px'
         }
       });
     }
@@ -1038,17 +1038,15 @@ var Button = function (_React$Component) {
       if (ripple) {
         rippleCls += ' mui--is-visible';
 
-        // css transform
-        var tCss = 'translate(-50%, -50%) translate(' + ripple.xPos + 'px,' + ripple.yPos + 'px)';
-
         // handle animation
-        if (ripple.isAnimating) rippleCls += ' mui--is-animating';else tCss = tCss + ' scale(0.0001, 0.0001)';
+        if (ripple.isAnimating) rippleCls += ' mui--is-animating';
 
         // style attrs
         rippleStyle = {
           width: ripple.diameter,
           height: ripple.diameter,
-          transform: tCss
+          top: ripple.top,
+          left: ripple.left
         };
       }
 
@@ -2851,7 +2849,8 @@ var Select = function (_React$Component) {
           defaultValue = _props.defaultValue,
           readOnly = _props.readOnly,
           useDefault = _props.useDefault,
-          reactProps = babelHelpers.objectWithoutProperties(_props, ['children', 'className', 'style', 'label', 'defaultValue', 'readOnly', 'useDefault']);
+          name = _props.name,
+          reactProps = babelHelpers.objectWithoutProperties(_props, ['children', 'className', 'style', 'label', 'defaultValue', 'readOnly', 'useDefault', 'name']);
 
 
       return _react2.default.createElement(
@@ -2868,6 +2867,7 @@ var Select = function (_React$Component) {
           'select',
           {
             ref: 'selectEl',
+            name: name,
             tabIndex: tabIndexInner,
             value: this.state.value,
             defaultValue: defaultValue,
@@ -2898,6 +2898,7 @@ var Select = function (_React$Component) {
 Select.propTypes = {
   label: PropTypes.string,
   value: PropTypes.string,
+  name: PropTypes.string,
   defaultValue: PropTypes.string,
   readOnly: PropTypes.bool,
   useDefault: PropTypes.bool,
@@ -2907,6 +2908,7 @@ Select.propTypes = {
 };
 Select.defaultProps = {
   className: '',
+  name: '',
   readOnly: false,
   useDefault: typeof document !== 'undefined' && 'ontouchstart' in document.documentElement ? true : false,
   onChange: null,