+++ packages/npm/lib/angular/button.js
@@ -19,14 +19,12 @@ var util = babelHelpers.interopRequireWildcard(_util);
 
 
 var moduleName = 'mui.button',
-    rippleClass = 'mui-ripple-effect',
     supportsTouch = 'ontouchstart' in document.documentElement,
     mouseDownEvents = supportsTouch ? 'touchstart' : 'mousedown',
-    mouseUpEvents = supportsTouch ? 'touchend' : 'mouseup mouseleave',
-    animationDuration = 600; /**
-                              * MUI Angular Button Component
-                              * @module angular/button
-                              */
+    mouseUpEvents = supportsTouch ? 'touchend' : 'mouseup mouseleave'; /**
+                                                                        * MUI Angular Button Component
+                                                                        * @module angular/button
+                                                                        */
 
 _angular2.default.module(moduleName, []).directive('muiButton', function () {
   return {
@@ -89,36 +87,27 @@ function mouseDownHandler(ev) {
   // get (x, y) position of click
   var offset = jqLite.offset(buttonEl),
       clickEv = ev.type === 'touchstart' ? ev.touches[0] : ev,
-      xPos = Math.round(clickEv.pageX - offset.left),
-      yPos = Math.round(clickEv.pageY - offset.top),
+      radius,
       diameter;
 
-  // calculate diameter
-  diameter = Math.sqrt(offset.width * offset.width + offset.height * offset.height) * 2 + 2 + 'px';
+  // calculate radius
+  radius = Math.sqrt(offset.width * offset.width + offset.height * offset.height);
 
-  // css transform
-  var tEnd = 'translate(-50%, -50%) translate(' + xPos + 'px,' + yPos + 'px)',
-      tStart = tEnd + ' scale(0.0001, 0.0001)';
+  diameter = radius * 2 + 'px';
 
+  // set position and dimensions
   jqLite.css(rippleEl, {
     width: diameter,
     height: diameter,
-    webkitTransform: tStart,
-    msTransform: tStart,
-    transform: tStart
+    top: Math.round(clickEv.pageY - offset.top - radius) + 'px',
+    left: Math.round(clickEv.pageX - offset.left - radius) + 'px'
   });
 
-  jqLite.addClass(rippleEl, 'mui--is-visible');
   jqLite.removeClass(rippleEl, 'mui--is-animating');
+  jqLite.addClass(rippleEl, 'mui--is-visible');
 
   // start animation
   util.requestAnimationFrame(function () {
-    jqLite.css(rippleEl, {
-      webkitTransform: tEnd,
-      msTransform: tEnd,
-      transform: tEnd
-    });
-
     jqLite.addClass(rippleEl, 'mui--is-animating');
   });
 }